Fragments now sends scan intent to service - fragment still doesn't receive the intents - fixed some comments - fixed folder structure
