+++ .idea/dictionaries/xeon.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="xeon">
+    <words>
+      <w>iparked</w>
+      <w>ocurance</w>
+    </words>
+  </dictionary>
+</component>
