+++ app/src/main/java/com/dsd2016/iparked_android/interfaces/OnImageButtonClicked.java
@@ -1,4 +1,4 @@
-package com.dsd2016.iparked_android.Interfaces;
+package com.dsd2016.iparked_android.interfaces;
 
 /**
  * Created by ricca on 23/11/2016.
