+++ app/src/main/java/com/dsd2016/iparked_android/myClasses/ClipRevealFrame.java
@@ -1,4 +1,4 @@
-package com.dsd2016.iparked_android.MyClasses;
+package com.dsd2016.iparked_android.myClasses;
 
 import android.content.Context;
 import android.graphics.Canvas;
