+++ app/src/main/java/com/dsd2016/iparked_android/services/OnBluetoothReceiver.java
@@ -1,4 +1,4 @@
-package com.dsd2016.iparked_android.Services;
+package com.dsd2016.iparked_android.services;
 
 import android.app.ActivityManager;
 import android.bluetooth.BluetoothAdapter;
