+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.dsd2016.iparked_android.Activities.MainActivity">
+    tools:context="com.dsd2016.iparked_android.activities.MainActivity">
     <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:id="@+id/fragment_container"
         android:layout_width="match_parent"
