+++ app/src/main/res/layout/activity_splash.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.dsd2016.iparked_android.Activities.SplashActivity">
+    tools:context="com.dsd2016.iparked_android.activities.SplashActivity">
 
     <ImageView
         android:layout_width="wrap_content"
