+++ app/src/main/res/layout/fragment_pairing.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.dsd2016.iparked_android.Fragments.MyPairingFragment"
+    tools:context="com.dsd2016.iparked_android.fragments.MyPairingFragment"
     android:id="@+id/root_layout">
 
     <ListView
@@ -33,7 +33,7 @@
 
     </android.support.design.widget.FloatingActionButton>
 
-    <com.dsd2016.iparked_android.MyClasses.ClipRevealFrame
+    <com.dsd2016.iparked_android.myClasses.ClipRevealFrame
         android:id="@+id/menu_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -87,5 +87,5 @@
             android:src="@drawable/find"
             android:tag="find" />
 
-    </com.dsd2016.iparked_android.MyClasses.ClipRevealFrame>
+    </com.dsd2016.iparked_android.myClasses.ClipRevealFrame>
 </FrameLayout>
