+++ app/src/main/res/layout/fragment_settings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.dsd2016.iparked_android.Fragments.MySettingsFragment"
+    tools:context="com.dsd2016.iparked_android.fragments.MySettingsFragment"
     android:id="@+id/root_layout">
 
     <ListView android:id="@android:id/list"
@@ -24,7 +24,7 @@
 
     </android.support.design.widget.FloatingActionButton>
 
-    <com.dsd2016.iparked_android.MyClasses.ClipRevealFrame
+    <com.dsd2016.iparked_android.myClasses.ClipRevealFrame
         android:id="@+id/menu_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -78,5 +78,5 @@
             android:src="@drawable/find"
             android:tag="find" />
 
-    </com.dsd2016.iparked_android.MyClasses.ClipRevealFrame>
+    </com.dsd2016.iparked_android.myClasses.ClipRevealFrame>
 </FrameLayout>
